Move tasks API requests into the `.Specification.TasksApi` namespace
