+++ src/OpenSearch.Client/Requests.Tasks.cs
@@ -41,7 +41,7 @@ using OpenSearch.Net.Specification.TasksApi;
 // ReSharper disable UnusedTypeParameter
 // ReSharper disable PartialMethodWithSinglePart
 // ReSharper disable RedundantNameQualifier
-namespace OpenSearch.Client
+namespace OpenSearch.Client.Specification.TasksApi
 {
 	[InterfaceDataContract]
 	public partial interface ICancelTasksRequest : IRequest<CancelTasksRequestParameters>
