+++ src/OpenSearch.Client/Tasks/CancelTasks/CancelTasksRequest.cs
@@ -26,26 +26,12 @@
 *  under the License.
 */
 
-using System.Threading.Tasks;
-using OpenSearch.OpenSearch.Xunit.XunitPlumbing;
-using OpenSearch.Client;
-using OpenSearch.Client.Specification.TasksApi;
-using Tests.Framework.EndpointTests;
-
-namespace Tests.Cluster.TaskManagement.GetTask
+namespace OpenSearch.Client.Specification.TasksApi
 {
-	public class GetTaskUrlTests : UrlTestsBase
-	{
-		[U] public override async Task Urls()
-		{
-			var id = new TaskId("fakeid:1");
+	[MapsApi("tasks.cancel.json")]
+	public partial interface ICancelTasksRequest { }
 
-			await UrlTester.GET($"/_tasks/fakeid%3A1")
-					.Fluent(c => c.Tasks.GetTask(id))
-					.Request(c => c.Tasks.GetTask(new GetTaskRequest(id)))
-					.FluentAsync(c => c.Tasks.GetTaskAsync(id))
-					.RequestAsync(c => c.Tasks.GetTaskAsync(new GetTaskRequest(id)))
-				;
-		}
-	}
+	public partial class CancelTasksRequest { }
+
+	public partial class CancelTasksDescriptor { }
 }
