+++ src/OpenSearch.Client/Tasks/GetTask/GetTaskRequest.cs
@@ -26,27 +26,12 @@
 *  under the License.
 */
 
-using System.Runtime.Serialization;
-using OpenSearch.Net;
-
-namespace OpenSearch.Client
+namespace OpenSearch.Client.Specification.TasksApi
 {
-	public class GetTaskResponse : ResponseBase
-	{
-		[DataMember(Name = "completed")]
-		public bool Completed { get; internal set; }
+	[MapsApi("tasks.get.json")]
+	public partial interface IGetTaskRequest { }
 
-		[DataMember(Name = "task")]
-		public TaskInfo Task { get; internal set; }
+	public partial class GetTaskRequest { }
 
-		[DataMember(Name = "response")]
-		internal LazyDocument Response { get; set; }
-
-		/// <summary>
-		/// Gets the response for the request that the task represents, if available.
-		/// Because the response will have no associated <see cref="ApiCallDetails"/>, the value
-		/// of <see cref="IResponse.IsValid"/> should not be used.
-		/// </summary>
-		public TResponse GetResponse<TResponse>() where TResponse : class, IResponse => Response?.AsUsingRequestResponseSerializer<TResponse>();
-	}
+	public partial class GetTaskDescriptor { }
 }
