+++ src/OpenSearch.Client/Tasks/ListTasks/ListTasksResponse.cs
@@ -31,7 +31,7 @@ using System.Runtime.Serialization;
 using OpenSearch.Net;
 using OpenSearch.Net.Utf8Json;
 
-namespace OpenSearch.Client
+namespace OpenSearch.Client.Specification.TasksApi
 {
 	[DataContract]
 	public class ListTasksResponse : ResponseBase
