+++ tests/Tests/Document/Multiple/DeleteByQueryRethrottle/DeleteByQueryRethrottleApiTests.cs
@@ -31,6 +31,7 @@ using System.Linq;
 using OpenSearch.Net;
 using FluentAssertions;
 using OpenSearch.Client;
+using OpenSearch.Client.Specification.TasksApi;
 using Tests.Core.Extensions;
 using Tests.Document.Multiple.Reindex;
 using Tests.Domain;
